pyproject.toml
[project]
name = "rag-hovedopgave"
version = "0.1.0"
description = "Curriculum-Aligned RAG Assistant for Datamatiker Students"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langsmith>=0.4.38",
    "chromadb>=1.3.0",
    "fastapi>=0.120.4",
    "gradio>=5.49.1",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "bs4>=0.0.2",
    "langchain-anthropic",
    "langchain-openai>=1.0.1",
    "langchain-chroma>=1.0.0",
    "pypdf>=6.1.3",
    "cohere>=5.20.0",
    "langchain-cohere>=0.5.0",
    "unstructured[md]>=0.18.15",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["F401", "F841"]  # F401: unused imports, F841: unused variables

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
